# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PathologIA",
        "description": "Documentation technique  de la plateforme PathologIA",
        "contact": {
            "email": "contact@pathologia.org"
        },
        "license": {
            "name": "Propriétaire"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/",
            "description": "Serveur de production 01"
        }
    ],
    "paths": {
        "/public/login": {
            "post": {
                "tags": [
                    "Connexion"
                ],
                "summary": "Authentification de l'utilisateur",
                "description": "Retourne un object utilisateur",
                "operationId": "login",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "Connexion"
                ],
                "summary": "Désauthentifie l'utilisateur connecté",
                "description": "Supprime la session de l'utilisateur",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/auth/admin/users": {
            "get": {
                "tags": [
                    "Admin/Utilisateur"
                ],
                "summary": "Affichage admin utilisateurs",
                "description": "Affiche le panneau d'administration des utilisateurs",
                "operationId": "create",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin/Utilisateur"
                ],
                "summary": "Création utilisateur",
                "description": "Créé un nouvel utilisateur",
                "operationId": "store",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/auth/user/me": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Affiche l'utilisateur",
                "description": "Affiche le profil de l'utilisateur connecté",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Mise à jour du profil",
                "description": "Mise à jour du profil de l'utilisateur connecté",
                "operationId": "update",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/public/password": {
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Réinitialisation password",
                "description": "Réinitialisation du mot de passe utilisateur",
                "operationId": "reset",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/auth/user/me/password": {
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Update password",
                "description": "Mise à jour du mot de passe utilisateur",
                "operationId": "updatePassword",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}